Providing an extra version without generator functions
