+++ no-generators.js
@@ -1,0 +1,3 @@
+const DB = require('./src/database')
+
+module.exports = DB
