+++ package.json
@@ -2,7 +2,7 @@
   "name": "sqlite3-helper",
   "version": "1.0.2",
   "description": "A wrapper library that eases the work with sqlite3 with some new functions and promises",
-  "main": "src/database.js",
+  "main": "src/generators.js",
   "scripts": {
     "test": "NODE_ENV=test ./node_modules/mocha/bin/_mocha --ui bdd --recursive ./test"
   },
