+++ src/database.js
@@ -147,20 +147,6 @@ DB.prototype.query = async function (query, ...bindParameters) {
 }
 
 /**
- * Similar to .query(), but instead of returning every row together, an iterator is returned so you can retrieve the rows one by one.
- * @see https://github.com/JoshuaWise/better-sqlite3/wiki/API#iteratebindparameters---iterator
- *
- * @param {Object} query the SQL-Query that should be run. Can contain placeholders for bind parameters.
- * @param {any} bindParameters You can specify bind parameters @see https://github.com/JoshuaWise/better-sqlite3/wiki/API#binding-parameters
- * @returns {Iterator}
- */
-DB.prototype.queryIterate = async function * (query, ...bindParameters) {
-  const statement = await this.prepare(query)
-  yield * statement.iterate(...bindParameters)
-  await statement.finalize()
-}
-
-/**
  * Returns the values of the first row of the query-result
  * @see https://github.com/JoshuaWise/better-sqlite3/wiki/API#getbindparameters---row
  *
